feat: Add list of UserValidator beans to UserConfig
This commit adds a list of UserValidator beans to the UserConfig class. These validators are intended to be used for validating user input in various scenarios.

Additionally, this commit updates the CreateCustomerUseCaseImpl class to inject the UserValidator beans as dependencies. This enables the use of these validators within the CreateCustomerUseCaseImpl for validating customer data before creation.